Time: 42 ms (86.31%) | Memory: 29.1 MB (47.49%) - LeetSync

+++ 15-3sum/3sum.cpp
@@ -10,7 +10,7 @@ public:
             if (nums[idx]>0) { break; }
 
             int left = idx+1;
-            int right = nums.size()-1;
+            int right = static_cast<int>(nums.size()-1);
             while (left<right)
             {
                 int leftValue = nums[left];
